+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "Noclip" effect; which shimmers the player so that they fall through blocks.
 - Added monolith-related effects; which temporarily change the background in-game.
 - Added "Switch Loadout" effect.
+- Added "Hiccup" effect.
 - Changed "Fart Effect" to spawn a poo item when the player is well-fed.
 - Fixed "Clear/Explode/Shuffle Inventory" effects not failing when there are no items.
 - Fixed "Screen Shake" effect not pausing when game is paused.